[so_long] mlx_beta swift 파일 수정

+++ so_long/mlx_beta/interface.swift
@@ -274,7 +274,7 @@ public func mlx_mouse_move_swift(_ winptr:UnsafeRawPointer, _ x:Int32, _ y:Int32
 ///	pt.y = sframe.size.y - frame.size.y - frame.origin.y + 1 + y
 	pt.y = frame.origin.y + frame.size.height - 1.0 - CGFloat(y)
 	CGWarpMouseCursorPosition(pt)
-	CGAssociateMouseAndMouseCursorPosition(UInt32(1))
+	CGAssociateMouseAndMouseCursorPosition(1)
 	return Int32(0);
 }
